requirements.txt
eel
pyserial
